move numba relted functions to a separated file
