+++ napari/utils/colormaps/colormap.py
@@ -16,6 +16,7 @@ from typing_extensions import Self
 
 from napari._pydantic_compat import Field, PrivateAttr, validator
 from napari.utils.color import ColorArray
+from napari.utils.colormaps import _colormap_numba as cmap_n
 from napari.utils.colormaps.colorbars import make_colorbar
 from napari.utils.colormaps.standardize_color import transform_color
 from napari.utils.compat import StrEnum
@@ -26,10 +27,6 @@ from napari.utils.translations import trans
 
 if TYPE_CHECKING:
     from numba import typed
-
-MAPPING_OF_UNKNOWN_VALUE = 0
-# For direct mode we map all unknown values to single value
-# for simplicity of implementation we select 0
 
 
 class ColormapInterpolationMode(StrEnum):
@@ -315,7 +312,7 @@ class CyclicLabelColormap(LabelColormapBase):
         return _cast_labels_data_to_texture_dtype_auto(values, self)
 
     def _map_without_cache(self, values) -> np.ndarray:
-        texture_dtype_values = _zero_preserving_modulo_numpy(
+        texture_dtype_values = cmap_n.zero_preserving_modulo_numpy(
             values,
             len(self.colors) - 1,
             values.dtype,
@@ -495,7 +492,7 @@ class DirectLabelColormap(LabelColormapBase):
         if mapper is not None:
             mapped = mapper[values]
         else:
-            values_cast = _labels_raw_to_texture_direct(values, self)
+            values_cast = cmap_n.labels_raw_to_texture_direct(values, self)
             mapped = self._map_precast(values_cast, apply_selection=True)
 
         if self.use_selection:
@@ -504,7 +501,7 @@ class DirectLabelColormap(LabelColormapBase):
 
     def _map_without_cache(self, values: np.ndarray) -> np.ndarray:
         cmap = self._cmap_without_selection()
-        cast = _labels_raw_to_texture_direct(values, cmap)
+        cast = cmap_n.labels_raw_to_texture_direct(values, cmap)
         return self._map_precast(cast, apply_selection=False)
 
     def _map_precast(self, values, apply_selection) -> np.ndarray:
@@ -582,10 +579,10 @@ class DirectLabelColormap(LabelColormapBase):
     ) -> tuple[dict[Optional[int], int], dict[int, np.ndarray]]:
         color_to_labels: dict[tuple[int, ...], list[Optional[int]]] = {}
         labels_to_new_labels: dict[Optional[int], int] = {
-            None: MAPPING_OF_UNKNOWN_VALUE
+            None: cmap_n.MAPPING_OF_UNKNOWN_VALUE
         }
         new_color_dict: dict[int, np.ndarray] = {
-            MAPPING_OF_UNKNOWN_VALUE: self.default_color,
+            cmap_n.MAPPING_OF_UNKNOWN_VALUE: self.default_color,
         }
 
         for label, color in self.color_dict.items():
@@ -633,7 +630,9 @@ class DirectLabelColormap(LabelColormapBase):
         from numba import typed, types
 
         # num_unique_colors + 2 because we need to map None and background
-        target_type = minimum_dtype_for_labels(self._num_unique_colors + 2)
+        target_type = cmap_n.minimum_dtype_for_labels(
+            self._num_unique_colors + 2
+        )
 
         dkt = typed.Dict.empty(
             key_type=getattr(types, data_dtype.name),
@@ -664,7 +663,7 @@ class DirectLabelColormap(LabelColormapBase):
                 'Cannot use numpy implementation for large values of labels '
                 'direct colormap. Please install numba.'
             )
-        dtype = minimum_dtype_for_labels(self._num_unique_colors + 2)
+        dtype = cmap_n.minimum_dtype_for_labels(self._num_unique_colors + 2)
         label_mapping = self._values_mapping_to_minimum_values_set()[0]
 
         # We need 2 + the max value: one because we will be indexing with the
@@ -672,7 +671,7 @@ class DirectLabelColormap(LabelColormapBase):
         # that index and map to the default value, rather than to the max
         # value in the map.
         mapper = np.full(
-            (max_value + 2), MAPPING_OF_UNKNOWN_VALUE, dtype=dtype
+            (max_value + 2), cmap_n.MAPPING_OF_UNKNOWN_VALUE, dtype=dtype
         )
         for key, val in label_mapping.items():
             if key is None:
@@ -776,14 +775,14 @@ def _cast_labels_data_to_texture_dtype_auto(
 
     data_arr = np.atleast_1d(data)
     num_colors = len(colormap.colors) - 1
-    zero_preserving_modulo_func = _zero_preserving_modulo
+    zero_preserving_modulo_func = cmap_n.zero_preserving_modulo
     if isinstance(data, np.integer):
-        zero_preserving_modulo_func = _zero_preserving_modulo_numpy
+        zero_preserving_modulo_func = cmap_n.zero_preserving_modulo_numpy
 
-    dtype = minimum_dtype_for_labels(num_colors + 1)
+    dtype = cmap_n.minimum_dtype_for_labels(num_colors + 1)
 
     if colormap.use_selection:
-        selection_in_texture = _zero_preserving_modulo_numpy(
+        selection_in_texture = cmap_n.zero_preserving_modulo_numpy(
             np.array([colormap.selection]), num_colors, dtype
         )
         converted = np.where(
@@ -798,103 +797,6 @@ def _cast_labels_data_to_texture_dtype_auto(
         return dtype.type(converted[0])
 
     return np.reshape(converted, original_shape)
-
-
-def _zero_preserving_modulo_numpy(
-    values: np.ndarray, n: int, dtype: np.dtype, to_zero: int = 0
-) -> np.ndarray:
-    """``(values - 1) % n + 1``, but with one specific value mapped to 0.
-
-    This ensures (1) an output value in [0, n] (inclusive), and (2) that
-    no nonzero values in the input are zero in the output, other than the
-    ``to_zero`` value.
-
-    Parameters
-    ----------
-    values : np.ndarray
-        The dividend of the modulo operator.
-    n : int
-        The divisor.
-    dtype : np.dtype
-        The desired dtype for the output array.
-    to_zero : int, optional
-        A specific value to map to 0. (By default, 0 itself.)
-
-    Returns
-    -------
-    np.ndarray
-        The result: 0 for the ``to_zero`` value, ``values % n + 1``
-        everywhere else.
-    """
-    res = ((values - 1) % n + 1).astype(dtype)
-    res[values == to_zero] = 0
-    return res
-
-
-def _zero_preserving_modulo_loop(
-    values: np.ndarray, n: int, dtype: np.dtype, to_zero: int = 0
-) -> np.ndarray:
-    """``(values - 1) % n + 1``, but with one specific value mapped to 0.
-
-    This ensures (1) an output value in [0, n] (inclusive), and (2) that
-    no nonzero values in the input are zero in the output, other than the
-    ``to_zero`` value.
-
-    Parameters
-    ----------
-    values : np.ndarray
-        The dividend of the modulo operator.
-    n : int
-        The divisor.
-    dtype : np.dtype
-        The desired dtype for the output array.
-    to_zero : int, optional
-        A specific value to map to 0. (By default, 0 itself.)
-
-    Returns
-    -------
-    np.ndarray
-        The result: 0 for the ``to_zero`` value, ``values % n + 1``
-        everywhere else.
-    """
-    result = np.empty_like(values, dtype=dtype)
-    # need to preallocate numpy array for asv memory benchmarks
-    return _zero_preserving_modulo_inner_loop(values, n, to_zero, out=result)
-
-
-def _zero_preserving_modulo_inner_loop(
-    values: np.ndarray, n: int, to_zero: int, out: np.ndarray
-) -> np.ndarray:
-    """``(values - 1) % n + 1``, but with one specific value mapped to 0.
-
-    This ensures (1) an output value in [0, n] (inclusive), and (2) that
-    no nonzero values in the input are zero in the output, other than the
-    ``to_zero`` value.
-
-    Parameters
-    ----------
-    values : np.ndarray
-        The dividend of the modulo operator.
-    n : int
-        The divisor.
-    to_zero : int
-        A specific value to map to 0. (Usually, 0 itself.)
-    out : np.ndarray
-        Preallocated output array
-
-    Returns
-    -------
-    np.ndarray
-        The result: 0 for the ``to_zero`` value, ``values % n + 1``
-        everywhere else.
-    """
-    for i in prange(values.size):
-        if values.flat[i] == to_zero:
-            out.flat[i] = 0
-        else:
-            out.flat[i] = (values.flat[i] - 1) % n + 1
-
-    return out
 
 
 @overload
@@ -950,87 +852,19 @@ def _cast_labels_data_to_texture_dtype_direct(
 
     if isinstance(data, np.integer):
         mapper = direct_colormap._label_mapping_and_color_dict[0]
-        target_dtype = minimum_dtype_for_labels(
+        target_dtype = cmap_n.minimum_dtype_for_labels(
             direct_colormap._num_unique_colors + 2
         )
         return target_dtype.type(
-            mapper.get(int(data), MAPPING_OF_UNKNOWN_VALUE)
+            mapper.get(int(data), cmap_n.MAPPING_OF_UNKNOWN_VALUE)
         )
 
     original_shape = np.shape(data)
     array_data = np.atleast_1d(data)
     return np.reshape(
-        _labels_raw_to_texture_direct(array_data, direct_colormap),
+        cmap_n.labels_raw_to_texture_direct(array_data, direct_colormap),
         original_shape,
     )
-
-
-def _labels_raw_to_texture_direct_numpy(
-    data: np.ndarray, direct_colormap: DirectLabelColormap
-) -> np.ndarray:
-    """Convert labels data to the data type used in the texture.
-
-    This implementation uses numpy vectorized operations.
-
-    See `_cast_labels_data_to_texture_dtype_direct` for more details.
-    """
-    if direct_colormap.use_selection:
-        return (data == direct_colormap.selection).astype(np.uint8)
-    mapper = direct_colormap._array_map
-    if any(x < 0 for x in direct_colormap.color_dict if x is not None):
-        half_shape = mapper.shape[0] // 2 - 1
-        data = np.clip(data, -half_shape, half_shape)
-    else:
-        data = np.clip(data, 0, mapper.shape[0] - 1)
-
-    return mapper[data]
-
-
-def _labels_raw_to_texture_direct_loop(
-    data: np.ndarray, direct_colormap: DirectLabelColormap
-) -> np.ndarray:
-    """
-    Cast direct labels to the minimum type.
-
-    Parameters
-    ----------
-    data : np.ndarray
-        The input data array.
-    direct_colormap : DirectLabelColormap
-        The direct colormap.
-
-    Returns
-    -------
-    np.ndarray
-        The cast data array.
-    """
-    if direct_colormap.use_selection:
-        return (data == direct_colormap.selection).astype(np.uint8)
-
-    dkt = direct_colormap._get_typed_dict_mapping(data.dtype)
-    target_dtype = minimum_dtype_for_labels(
-        direct_colormap._num_unique_colors + 2
-    )
-    result_array = np.full_like(
-        data, MAPPING_OF_UNKNOWN_VALUE, dtype=target_dtype
-    )
-    return _labels_raw_to_texture_direct_inner_loop(data, dkt, result_array)
-
-
-def _labels_raw_to_texture_direct_inner_loop(
-    data: np.ndarray, dkt: 'typed.Dict', out: np.ndarray
-) -> np.ndarray:
-    """
-    Relabel data using typed dict with mapping unknown labels to default value
-    """
-    # The numba typed dict does not provide official Api for
-    # determine key and value types
-    for i in prange(data.size):
-        val = data.flat[i]
-        if val in dkt:
-            out.flat[i] = dkt[data.flat[i]]
-
-    return out
 
 
 def _texture_dtype(num_colors: int, dtype: np.dtype) -> np.dtype:
@@ -1045,47 +879,7 @@ def _texture_dtype(num_colors: int, dtype: np.dtype) -> np.dtype:
         return np.dtype(np.uint8)
     if dtype.itemsize == 2:
         return np.dtype(np.uint16)
-    return minimum_dtype_for_labels(num_colors)
-
-
-def minimum_dtype_for_labels(num_colors: int) -> np.dtype:
-    """Return the minimum texture dtype that can hold given number of colors.
-
-    Parameters
-    ----------
-    num_colors : int
-        Number of unique colors in the data.
-
-    Returns
-    -------
-    np.dtype
-        Minimum dtype that can hold the number of colors.
-    """
-    if num_colors <= np.iinfo(np.uint8).max:
-        return np.dtype(np.uint8)
-    if num_colors <= np.iinfo(np.uint16).max:
-        return np.dtype(np.uint16)
-    return np.dtype(np.float32)
-
-
-try:
-    import numba
-except ModuleNotFoundError:
-    _zero_preserving_modulo = _zero_preserving_modulo_numpy
-    _labels_raw_to_texture_direct = _labels_raw_to_texture_direct_numpy
-    prange = range
-else:
-    _zero_preserving_modulo_inner_loop = numba.njit(parallel=True, cache=True)(
-        _zero_preserving_modulo_inner_loop
-    )
-    _zero_preserving_modulo = _zero_preserving_modulo_loop
-    _labels_raw_to_texture_direct = _labels_raw_to_texture_direct_loop
-    _labels_raw_to_texture_direct_inner_loop = numba.njit(
-        parallel=True, cache=True
-    )(_labels_raw_to_texture_direct_inner_loop)
-    prange = numba.prange  # type: ignore [misc]
-
-    del numba
+    return cmap_n.minimum_dtype_for_labels(num_colors)
 
 
 def _normalize_label_colormap(
